+++ cedar/devices/Locomotion.cpp
@@ -45,14 +45,29 @@
 
 // SYSTEM INCLUDES
 
+
+//------------------------------------------------------------------------------
+// static variables
+//------------------------------------------------------------------------------
+
+const cedar::dev::Component::ComponentDataType cedar::dev::Locomotion::FORWARD_VELOCITY = 1;
+const cedar::dev::Component::ComponentDataType cedar::dev::Locomotion::TURNING_RATE = 2;
 //----------------------------------------------------------------------------------------------------------------------
 // constructors and destructor
 //----------------------------------------------------------------------------------------------------------------------
 
+void cedar::dev::Locomotion::init()
+{
+  installCommandAndMeasurementType( cedar::dev::Locomotion::FORWARD_VELOCITY, "Forward Velocity" );
+  installCommandAndMeasurementType( cedar::dev::Locomotion::TURNING_RATE, "Turning Rate" );
+
+  setCommandAndMeasurementDimensionality( cedar::dev::Locomotion::FORWARD_VELOCITY, 1 );
+  setCommandAndMeasurementDimensionality( cedar::dev::Locomotion::TURNING_RATE, 1 );
+
+}
+
 cedar::dev::Locomotion::Locomotion()
 :
-mForwardVelocity(new cedar::aux::VelocityData(0.0 * cedar::unit::meters_per_second)),
-mTurningRate(new cedar::aux::AngularVelocityData(0.0 * cedar::unit::radians_per_second)),
 _mForwardVelocityLimits
 (
   new cedar::aux::math::VelocityLimitsParameter
@@ -82,16 +97,13 @@ _mTurningRateLimits
       )
 )
 {
-  this->addCommandedData("turning rate", this->mTurningRate, boost::bind(&cedar::dev::Locomotion::applyTurningRate, this));
-  this->addCommandedData("forward velocity", this->mForwardVelocity, boost::bind(&cedar::dev::Locomotion::applyTurningRate, this));
+  init();
 }
 
 // constructor taking an externally created channel
 cedar::dev::Locomotion::Locomotion(cedar::dev::ChannelPtr channel)
 :
 cedar::dev::Component(channel),
-mForwardVelocity(new cedar::aux::VelocityData(0.0 * cedar::unit::meters_per_second)),
-mTurningRate(new cedar::aux::AngularVelocityData(0.0 * cedar::unit::radians_per_second)),
 _mForwardVelocityLimits
 (
   new cedar::aux::math::VelocityLimitsParameter
@@ -121,33 +133,73 @@ _mTurningRateLimits
       )
 )
 {
-  //!@todo It is inefficient to call two different functions for setting turning rate and velocity, there should be only one call.
-  this->addCommandedData("turning rate", this->mTurningRate, boost::bind(&cedar::dev::Locomotion::applyTurningRate, this));
-  this->addCommandedData("forward velocity", this->mForwardVelocity, boost::bind(&cedar::dev::Locomotion::applyTurningRate, this));
+  init();
 }
 
 //----------------------------------------------------------------------------------------------------------------------
 // methods
 //----------------------------------------------------------------------------------------------------------------------
-
-void cedar::dev::Locomotion::applyTurningRate()
+void cedar::dev::Locomotion::setForwardVelocity(cedar::unit::Velocity forwardVelocity)
 {
-  this->setTurningRate(this->mTurningRate->getData());
+  setForwardVelocityAndTurningRate(forwardVelocity, getTurningRate());
 }
 
-void cedar::dev::Locomotion::applyForwardVelocity()
+void cedar::dev::Locomotion::setTurningRate(cedar::unit::AngularVelocity turningRate)
 {
-  this->setForwardVelocity(this->mForwardVelocity->getData());
+  setForwardVelocityAndTurningRate(getForwardVelocity(), turningRate);
+}
+
+void cedar::dev::Locomotion::setForwardVelocityAndTurningRate
+     (
+       cedar::unit::Velocity forwardVelocity,
+       cedar::unit::AngularVelocity turningRate
+     )
+{
+  cv::Mat forward_matrix = cv::Mat(1,1,CV_32F);
+  cv::Mat turning_matrix = cv::Mat(1,1,CV_32F);
+
+  //@todo: use a generic limit mechanism
+  thresholdForwardVelocity(forwardVelocity);
+  thresholdTurningRate(turningRate);
+
+  forward_matrix.at<float>(0,0) = forwardVelocity / cedar::unit::DEFAULT_VELOCITY_UNIT;
+  turning_matrix.at<float>(0,0) = turningRate / cedar::unit::DEFAULT_ANGULAR_VELOCITY_UNIT;
+
+  setUserSideCommandBuffer( cedar::dev::Locomotion::FORWARD_VELOCITY, forward_matrix);
+  setUserSideCommandBuffer( cedar::dev::Locomotion::TURNING_RATE, turning_matrix);
+
+/*
+  cedar::unit::Velocity left_wheel_speed = 0.0 * cedar::unit::DEFAULT_VELOCITY_UNIT;
+  cedar::unit::Velocity right_wheel_speed = 0.0 * cedar::unit::DEFAULT_VELOCITY_UNIT;
+  convertToWheelSpeed(forwardVelocity, turningRate, left_wheel_speed, right_wheel_speed);
+
+  thresholdToHardwareLimits(left_wheel_speed, right_wheel_speed);
+
+  mWheelSpeed[0] = left_wheel_speed;
+  mWheelSpeed[1] = right_wheel_speed;
+
+  cedar::unit::AngularVelocity turning_rate;
+  cedar::unit::Velocity forward_velocity;
+  convertToForwardVelocityAndTurningRate(left_wheel_speed, right_wheel_speed, forward_velocity, turning_rate);
+  this->mTurningRate->setData(turning_rate);
+  this->mForwardVelocity->setData(forward_velocity);
+
+  sendMovementCommand();
+*/
 }
 
 cedar::unit::Velocity cedar::dev::Locomotion::getForwardVelocity() const
 {
-  return mForwardVelocity->getData();
+  cv::Mat ret = getUserSideMeasurementBuffer( cedar::dev::Locomotion::FORWARD_VELOCITY );
+
+  return double(ret.at<float>(0,0)) * cedar::unit::DEFAULT_VELOCITY_UNIT;
 }
 
 cedar::unit::AngularVelocity cedar::dev::Locomotion::getTurningRate() const
 {
-  return mTurningRate->getData();
+  cv::Mat ret = getUserSideMeasurementBuffer( cedar::dev::Locomotion::TURNING_RATE );
+
+  return double(ret.at<float>(0,0)) * cedar::unit::DEFAULT_ANGULAR_VELOCITY_UNIT;
 }
 
 void cedar::dev::Locomotion::thresholdForwardVelocity(cedar::unit::Velocity& forwardVelocity) const
@@ -168,3 +220,4 @@ void cedar::dev::Locomotion::stop()
     0.0 * cedar::unit::DEFAULT_ANGULAR_VELOCITY_UNIT
   );
 }
+