+++ cedar/devices/NetworkChannel.cpp
@@ -50,6 +50,11 @@ _mPort(new cedar::aux::UIntParameter(this, "port", 10000, 0, 65535))
 {
 }
 
+cedar::dev::NetworkChannel::~NetworkChannel()
+{
+  cedar::dev::Channel::prepareChannelDestructAbsolutelyRequired();
+}
+
 //----------------------------------------------------------------------------------------------------------------------
 // methods
 //----------------------------------------------------------------------------------------------------------------------