+++ cedar/processing/gui/Group.h
@@ -65,6 +65,9 @@
 #include <string>
 #include <list>
 
+#include "cedar/auxiliaries/gui/Viewer.h"
+#include "cedar/devices/gui/KinematicChainWidget.h"
+
 /*!@brief The representation of a cedar::proc::Group in a cedar::proc::gui::Scene.
  *
  *        This class takes care of loading cedar::proc::Groups in a manner that allows them to be added into
@@ -179,6 +182,13 @@ public:
   {
     return this->mpScene;
   }
+
+  //! Opens the kinematic chain control widget from saved configuration file
+  void openKinematicChainWidget(const aux::ConfigurationNode &node);
+
+  //! Opens a view on the simulated scene
+  void openSceneViewer();
+  void openSceneViewer(const cedar::aux::ConfigurationNode& node);
 
   /*!@brief Sets the ui configuration for the element when it is added to the group.
    */
@@ -458,6 +468,10 @@ private slots:
 
   void geometryLockChanged();
 
+  void removeViewer();
+
+  void removeKinematicChainWidget();
+
   //--------------------------------------------------------------------------------------------------------------------
   // members
   //--------------------------------------------------------------------------------------------------------------------
@@ -472,6 +486,12 @@ private:
 
   //!@brief a filename from which to load a group configuration, or to which to save a configuration
   mutable std::string mFileName;
+
+  //!@brief a vector containing all the viewers
+  std::vector<cedar::aux::gui::Viewer*> mViewers;
+
+  //!@brief a vector containing all the KinematicChainWidget for savin purposes...
+  std::vector<cedar::dev::gui::KinematicChainWidget*> mKinematicChainWidgets;
 
   //!@brief a vector of all source connectors
   std::vector<cedar::proc::gui::DataSlotItem*> mConnectorSources;