+++ tests/interactive/devices/gl/sdh/main.cpp
@@ -51,10 +51,10 @@
 int main(int argc, char **argv)
 {
   // find resources
-  std::string finger_one_configuration_file = cedar::aux::locateResource("configs/sdh_finger_one.json");
-  std::string finger_two_configuration_file = cedar::aux::locateResource("configs/sdh_finger_two.json");
-  std::string finger_three_configuration_file = cedar::aux::locateResource("configs/sdh_finger_three.json");
-  std::string palm_configuration_file = cedar::aux::locateResource("configs/sdh_palm.json");
+  std::string finger_one_configuration_file = cedar::aux::locateResource("robots/sdh_finger_one.json");
+  std::string finger_two_configuration_file = cedar::aux::locateResource("robots/sdh_finger_two.json");
+  std::string finger_three_configuration_file = cedar::aux::locateResource("robots/sdh_finger_three.json");
+  std::string palm_configuration_file = cedar::aux::locateResource("robots/sdh_palm.json");
 
   QApplication a(argc, argv);
 
@@ -67,7 +67,7 @@ int main(int argc, char **argv)
   finger_three->readJson(finger_three_configuration_file);
   cedar::dev::KinematicChainPtr palm(new cedar::dev::SimulatedKinematicChain());
   palm->readJson(palm_configuration_file);
-//  finger_one->getRootCoordinateFrame()->setTransformation(cv::Mat::eye(4, 4, CV_64FC1));
+//  finger_one->getRootCoordinateFrame()->setTransformation(cv::Mat::eye(4, 4, CV_32FC1));
 
   // create gl visualization objects
   cedar::dev::gl::SdhPtr hand_visualization